# maintence-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maintence API",
        "description": "API Documentation for Maintenance System",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send Reset Link",
                "description": "Generates a reset token and logs it (or sends email). Check storage/logs/laravel.log for the link.",
                "operationId": "3b968e5662bb656a437a1b0b32e13f02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@kawarem.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset Password",
                "description": "Resets the user password using the token. NOTE: Stores password AS IS (Client-side hashing assumed).",
                "operationId": "5a1e02ae971927f2615b5ff0e3f26f0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "69e2c6c7..."
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@kawarem.com"
                                    },
                                    "password": {
                                        "description": "The ALREADY HASHED password from client",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password has been reset"
                    },
                    "400": {
                        "description": "Invalid token or email"
                    }
                }
            }
        },
        "/api/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password (Logged In)",
                "description": "Updates the user password. Requires the current password for verification.",
                "operationId": "426aa8e2e128ccd6a0e78584373681b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "old_hashed_password_from_client"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "new_hashed_password_from_client"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully"
                    },
                    "401": {
                        "description": "Incorrect current password"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/record-parts": {
            "get": {
                "tags": [
                    "Inventory Records"
                ],
                "summary": "Get Records (Chunked 10)",
                "description": "Returns records in chunks of 10. Default is Page 1 (Records 1-10). Send ?page=2 for 11-20.",
                "operationId": "284c8853d51408287adc77830b58f068",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Chunk number (Default: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of records",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RecordPartResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory Records"
                ],
                "summary": "Add Stock Record",
                "description": "Adds a record and updates Part stock. Type 1 increases stock; 2,3,4 decrease it.",
                "operationId": "62ca1c16150b7dc6dc56968cec2862d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "part_id",
                                    "modify_type_id",
                                    "modified_amount",
                                    "modification_date"
                                ],
                                "properties": {
                                    "part_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "modify_type_id": {
                                        "description": "1=Add, 2=Consumed, 3=Wasted, 4=Refunded",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "modified_amount": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "modification_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-06"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/record-parts/{id}": {
            "get": {
                "tags": [
                    "Inventory Records"
                ],
                "summary": "Get Record Details",
                "operationId": "ed1bbe751da186616f6d97ff6f04680e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RecordPartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventory Records"
                ],
                "summary": "Update Record & Re-calculate Stock",
                "description": "Updates the record and automatically adjusts the part stock based on the difference.",
                "operationId": "b07221746516ff33f507f9e730ad295d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "modified_amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "RecordPartResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "part_name": {
                        "type": "string",
                        "example": "Compressor"
                    },
                    "modify_type": {
                        "type": "string",
                        "example": "Restock"
                    },
                    "amount": {
                        "type": "integer",
                        "example": 10
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}